rename filename and fix typo
